[AutoDiff] Move ActivityAnalysis to a separate file.
